Add new access mode to represent address register modification.
Also modify extract_addr_expr so that it stops the register expansion
when the resulting expression gets too complicated. Instead, the
unexpanded register is used as the base reg, and a reg-mod access is
placed into the access sequence that stores the expression the register
is set to inside m_reg_mod_expr.