Restructure the reg_mod type access.
The modified register is now stored in m_mod_reg.  The address that
it's set to is in m_addr_expr, or m_addr_rtx if it doesn't match
the addr_expr format.
Also add an m_cost and m_used field that have the same function as in
the reg_values class.
With the access class modified this way, it can be used in
update_insn_stream instead of reg_values.