Record the original register modifying insns in the access sequence.
Add the address modifying insns of the original insn list to the
sequence as reg_mod accesses.  These are marked as "removable" so that
they are removed in gen_address_mod when new address modifications are
generated.
This will make it easier to compare the costs of address calculation in
the original insn stream and in modified versions.