Start the search from the given insn in find_reg_value.
This allows to revert the changes in 3e4a2201e554, which caused problems
when the access' insn was the last insn in the last BB.
Also adjust extract_addr_expr and the places where it's called so that
the search starts from the correct place in those cases.
Also change some of the comments and parameter names in extract_addr_expr.