Make the emit_before/after_insn constraints less strict.
Instead of getting inserted right before/after the specified insn, the
reg_mod's insn can be inserted normally in the sequence as long as that
makes it end up before/after the specified insn.

The only case where the reg_mod insn should go right after the specified
insn is when cloning regs that get clobbered or set to an unknown value.
In these cases, we can simply leave in the clobbering access and place
the cloning access after it, which will make update_insn_stream emit the
insns to the correct place without any emit_after constraints.