increase costs for stack-frame related accesses
for now, we can't run AMS after register allocation and stack-frame
layout.  because the stack-frame layout is not known when AMS is ran, it
can't make good choices.  it seems to be better to avoid reg mods and
auto-mod modes on stack-frame related accsess as these tend to use more
live registers and increase register pressure.