Don't skip the last insn of a basic block when tracing back reg values across BBs.
This required rewriting rtx_to_addr_expr so that the search is started
from curr_insn, not the insn before it. Since curr_insn can be null in
this case (e.g. when calling with
curr_insn=prev_nonnote_insn_bb (mod_insn)), the current BB needs to be
stored separately.