Insert the reg-reg copies to the right place.
If a reg's previous value is reused, the reg copy that stores its value
should go after the reg-mod where it was set to that value.
This change required changing the type of the starting addresses from
reg_mod* to sequence::iterator so that new elements can be inserted
after them.